Added FARS COA Terms
I originally added these in a branch on a fork of the repo but for some reason the GitHub interface does not allow me to declare a PR to the primary project. Maybe it's a write access issue? I have added these directly but if you decide to reject these I may need some help finding a work around to the PR issue. The NTR request is here: https://github.com/obi-ontology/obi/issues/1637